modified User model, fixed privacy issue user routes, modified product routes

+++ routes/product.routes.js
@@ -5,7 +5,7 @@ const { isAuthenticated } = require("../middlewares/auth.middleware")
 
 // Endpoint / => /api/products
 
-// GET - Reads all products
+// GET all products
 router.get("/", async (req, res) => {
     try {
       const allProducts = await Product.find();
@@ -16,7 +16,7 @@ router.get("/", async (req, res) => {
     }
   });
   
-// GET - Reads one product & populates with reviews
+// GET product and populate with reviews
 router.get("/:productId", async (req, res) => {
   const { productId } = req.params;
   try {
@@ -32,27 +32,29 @@ router.get("/:productId", async (req, res) => {
   }
 })
 
-// POST - Creates one review associated with one product
+// POST review associated with one product
 router.post("/:productId/reviews", isAuthenticated, async (req, res) => {
   const { productId } = req.params;
-  const payload = req.body;
   const { userId } = req.tokenPayload;
-  payload.author = userId; // sets the author to the authenticated user's ID
+  const payload = req.body;
+  payload.product = productId; // Associates the review with the product
+  payload.author = userId; // Sets the author to the authenticated user's ID
 
   try {
-    const product = await Product.findById(productId);
+    // Retrieves the book form the database using its ID (to ensure the book exists and can be updated)
+    const product = await Product.findById(productId); 
     if (!product) {
       return res.status(404).json({ message: "Product not found" })
     }
-    payload.product = productId; // Associate the review with the product
 
+    // creates a review
     const review = await Review.create(payload);
     if (!review) {
       return res.status(404).json({ message: "Review not found" })
     }
 
     // Push the review directly into the product's reviews array
-    await Product.findByIdAndUpdate(productId, { $push: { reviews: review } }, { new: true});
+    await Product.findByIdAndUpdate(productId, { $push: { reviews: review._id } }, { new: true});
     res.status(201).json({ message: "Review created successfully", review });
 
   } catch (error) {
@@ -61,17 +63,34 @@ router.post("/:productId/reviews", isAuthenticated, async (req, res) => {
   }
 });
 
-// PUT - Updates one review
+// PUT (update) review 
 router.put("/:productId/reviews/:reviewId", isAuthenticated, async (req, res) => {
-  const { reviewId } = req.params;
+  const { productId, reviewId } = req.params;
+  const { userId } = req.tokenPayload; 
   const payload = req.body;
 
   try {
-    const updatedReview = await Review.findByIdAndUpdate(reviewId, req.body, { new: true});
-    if (!updatedReview) {
-      return res.status(404).json({ message: "Updated review not found" })
-    }
-    res.status(200).json(updatedReview);
+    // Checks if review exists in the database
+    const review = await Review.findById(reviewId)
+      if (!review) {
+        return res.status(404).json({ message: "Review not found."})
+      }
+
+      // Checks if the review belongs to the specified product
+      if (review.product !== productId) {
+        return res.status(400).json({ message: "Review does not belong to the specified product." });
+      }
+
+      if (review.author == userId) {
+      // Checks if the review exists in the database and updates it
+      const updatedReview = await Review.findByIdAndUpdate(reviewId, payload, { new: true});
+      res.status(200).json(updatedReview);
+        if (!updatedReview) {
+          return res.status(404).json({ message: "Updated review not found." })
+        } else {
+          res.status(403).json({message: "You are not the right user to update this review."});
+        }
+      }
 
   } catch (error) {
     console.log(error);
@@ -79,16 +98,36 @@ router.put("/:productId/reviews/:reviewId", isAuthenticated, async (req, res) =>
   }
 })
 
-// DELETE -  Deletes one review
+// DELETE review
 router.delete("/:productId/reviews/:reviewId", isAuthenticated, async (req, res) => {
-  const { reviewId } = req.params
+  const { productId, reviewId } = req.params
+  const { userId } = req.tokenPayload; 
+
    try {
-    const deletedReview = await Review.findByIdAndDelete(reviewId);
-    if (!deletedReview) {
-      return res.status(404).json({ message: "Deleted review not found or already deleted." })
+    // Checks if the review exists in the database 
+    const review = await Review.findById(reviewId)
+    if (!review) {
+      return res.status(404).json({ message: "Review not found."})
     }
-    res.status(200).send();
-    //res.status(200).json({ messafe: "Review deleted successfully."})
+
+    // Checks if the review belongs to the specified product
+    if (review.product !== productId) { // review.product.toString() !== productId ?
+      return res.status(400).json({ message: "Review does not belong to the specified product." });
+    }
+
+    // Checks if the user is allowed to delete the review
+    if (review.author == userId) {
+      // Checks if the review exists in the database and deletes it
+      const deletedReview = await Review.findByIdAndDelete(reviewId);
+      if (!deletedReview) {
+        return res.status(404).json({ message: "Deleted review not found or already deleted." })
+      }
+      res.status(200).send(); // Sends a plain text response to the client
+      //res.status(200).json({ message: "Review deleted successfully."}) // Sends a JSON response with structured data included. 
+    } else {
+      res.status(400).json({ message: "You are not allowed to delete the review." });
+    }
+
   } catch (error) {
     res.status(500).json({ error, message: "Failed to delete review." });
   }
